Refactor : 파일 생성 로직 다른 파일로 분리
